adds `pathlib.Path` objects as `PathLike`
Currently, the (local to Partitura) `PathLike` check accepts only `str`, `bytes`, and `os.PathLike`.

This PR proposes the addition of `pathlib.Path` objects.

The consequence of non-inclusion in this `PathLike` list is attempting handling as an URL (`io/__init__`) which leads to less-than-ideal error messages. 

This PR is limited to the case of `pathlib.Path` objects; in another PR it might be worth additional work on the modular handling of further types and/or the real case of URLs.